Use sentimentr package
sentimentr package already does a lot of the heavy lifting of sentiment prediction.

The situations where it gets wrong tend to be different, based on the TRUE sentiment: False Negatives tend to be longer in length, and False Positives tend to be shorter in length.

Therefore, just try a model with the predicted sentiment, sentiment turned into binary, and word length.